requirements.txt
auto_gpt_plugin_template
black
isort
flake8
finx-io
numpy
pandas
pathlib
python-dotenv
requests
ta

